Fix build - Bad forward declared underlying type on enum
#jira UE-108513
#rb trivial

[CL 15406663 by benoit deschenes in Main branch]

+++ Engine/Plugins/Enterprise/DatasmithImporter/Source/DatasmithImporter/Public/DatasmithMaterialExpressions.h
@@ -33,7 +33,7 @@ class UMaterialInterface;
 struct FExpressionInput;
 class UMaterialExpressionMaterialFunctionCall;
 
-enum class EDatasmithMaterialExpressionType : uint64;
+enum class EDatasmithMaterialExpressionType : uint8;
 enum class EDatasmithTextureMode : uint8;
 
 enum class EDatasmithTextureSlot
